# apps/simulator/openapi.yaml
openapi: 3.0.3
info:
  title: Moveet Simulator API
  description: |
    REST API for the Moveet vehicle fleet simulation engine.
    Simulates vehicle movement on real road networks (OpenStreetMap) with
    A* pathfinding, heat zones, incidents, recording/replay, and fleet management.
  version: 0.0.4
  license:
    name: MIT
servers:
  - url: http://localhost:5010
    description: Local development server
...
paths:
...
  /vehicles:
    get:
      operationId: getVehicles
      summary: List all vehicles with current state
      tags: [Vehicles]
      responses:
        "200":
          description: Array of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VehicleDTO"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /directions:
    get:
      operationId: getDirections
      summary: Get all active vehicle directions/routes
      tags: [Vehicles]
      responses:
        "200":
          description: Array of active directions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Direction"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /direction:
    post:
      operationId: setDirections
      summary: Set destinations for one or more vehicles
      tags: [Vehicles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              minItems: 1
              items:
                $ref: "#/components/schemas/DirectionRequest"
      responses:
        "200":
          description: Direction results
          content:
            application/json:
              schema:
                type: object
                required: [status, results]
                properties:
                  status:
                    type: string
                    enum: [direction]
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/DirectionResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"

  # ─── Network ─────────────────────────────────────────────────────────
...
  /find-node:
    post:
      operationId: findNode
      summary: Find the nearest graph node to given coordinates
      tags: [Network]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoordinatePair"
      responses:
        "200":
          description: Snapped coordinates [longitude, latitude]
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoordinatePair"
        "400":
          description: Invalid coordinates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /find-road:
    post:
      operationId: findRoad
      summary: Find the nearest road to given coordinates
      tags: [Network]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoordinatePair"
      responses:
        "200":
          description: Nearest road object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Road"
        "400":
          description: Invalid coordinates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /search:
    post:
      operationId: searchByName
      summary: Search roads and POIs by name
      tags: [Network]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                  minLength: 1
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchResult"
        "400":
          description: Invalid query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ─── Heat Zones ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ValidationError:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: string
    CoordinatePair:
      type: array
      items:
        type: number
      minItems: 2
      maxItems: 2
      description: "[longitude, latitude]"
      example: [36.8219, -1.2921]
    LatLngPair:
      type: array
      items:
        type: number
      minItems: 2
      maxItems: 2
      description: "[latitude, longitude]"

    # ─── Vehicle types ───────────────────────────────────────────────
    VehicleType:
      type: string
      enum: [car, truck, motorcycle, ambulance, bus]
...
    HighwayType:
      type: string
      enum: [motorway, trunk, primary, secondary, tertiary, residential]
...
    VehicleDTO:
      type: object
      required: [id, name, type, position, speed, heading]
      properties:
        id:
          type: string
        name:
          type: string
        type:
          $ref: "#/components/schemas/VehicleType"
        position:
          $ref: "#/components/schemas/LatLngPair"
        speed:
          type: number
          description: Current speed in km/h
        heading:
          type: number
          description: Bearing in degrees (0-360)
        fleetId:
          type: string

    # ─── Simulation ──────────────────────────────────────────────────
...
    DirectionRequest:
      type: object
      required: [id]
      properties:
        id:
          type: string
          description: Vehicle ID
        lat:
          type: number
          description: Destination latitude (used when waypoints not provided)
        lng:
          type: number
          description: Destination longitude (used when waypoints not provided)
        waypoints:
          type: array
          items:
            $ref: "#/components/schemas/WaypointRequest"
          description: Multi-stop waypoints (overrides lat/lng if provided)
    WaypointRequest:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
        lng:
          type: number
        dwellTime:
          type: number
          description: Seconds to dwell at this waypoint
        label:
          type: string
    Direction:
      type: object
      required: [vehicleId, route]
      properties:
        vehicleId:
          type: string
        route:
          $ref: "#/components/schemas/Route"
        eta:
          type: number
          description: Estimated time of arrival in seconds
        waypoints:
          type: array
          items:
            $ref: "#/components/schemas/Waypoint"
        currentWaypointIndex:
          type: number
    DirectionResult:
      type: object
      required: [vehicleId, status]
      properties:
        vehicleId:
          type: string
        status:
          type: string
          enum: [ok, error]
        error:
          type: string
        route:
          type: object
          properties:
            start:
              $ref: "#/components/schemas/LatLngPair"
            end:
              $ref: "#/components/schemas/LatLngPair"
            distance:
              type: number
        eta:
          type: number
        snappedTo:
          $ref: "#/components/schemas/LatLngPair"
        waypointCount:
          type: number
        legs:
          type: array
          items:
            type: object
            properties:
              start:
                $ref: "#/components/schemas/LatLngPair"
              end:
                $ref: "#/components/schemas/LatLngPair"
              distance:
                type: number
    Route:
      type: object
      required: [edges, distance]
      properties:
        edges:
          type: array
          items:
            $ref: "#/components/schemas/RouteEdge"
        distance:
          type: number
          description: Total route distance in km
    RouteEdge:
      type: object
      required: [id, streetId, start, end, distance, bearing, highway, maxSpeed, surface, oneway]
      properties:
        id:
          type: string
        streetId:
          type: string
        name:
          type: string
        start:
          type: object
          required: [id, coordinates]
          properties:
            id:
              type: string
            coordinates:
              $ref: "#/components/schemas/LatLngPair"
        end:
          type: object
          required: [id, coordinates]
          properties:
            id:
              type: string
            coordinates:
              $ref: "#/components/schemas/LatLngPair"
        distance:
          type: number
        bearing:
          type: number
        highway:
          $ref: "#/components/schemas/HighwayType"
        maxSpeed:
          type: number
        surface:
          type: string
        oneway:
          type: boolean
    Waypoint:
      type: object
      required: [position]
      properties:
        position:
          $ref: "#/components/schemas/LatLngPair"
        dwellTime:
          type: number
        label:
          type: string

    # ─── Network ─────────────────────────────────────────────────────
...
    Road:
      type: object
      required: [name, nodeIds, streets]
      properties:
        name:
          type: string
        nodeIds:
          description: Set of node IDs serialized as array
          oneOf:
            - type: array
              items:
                type: string
            - type: object
        streets:
          type: array
          items:
            type: array
            items:
              $ref: "#/components/schemas/CoordinatePair"
...
    SearchResult:
      type: object
      required: [name, nodeIds, coordinates]
      properties:
        name:
          type: string
        nodeIds:
          type: array
          items:
            type: string
        coordinates:
          type: array
          items:
            $ref: "#/components/schemas/LatLngPair"

    # ─── Heat Zones ──────────────────────────────────────────────────
...